[BE-2001] feat: 선생님 문제집 목록 API

+++ forest-back/forest-workbook/src/main/java/com/ssafy/forestworkbook/dto/workbook/TeacherWorkbookDto.java
@@ -7,7 +7,7 @@ import lombok.*;
 @ToString
 public class TeacherWorkbookDto {
 
-    private int workbookId;
+    private Long workbookId;
     private String title;
     private String workbookImgPath;
     private int bookmarkCount;
